[qca-ssdk]Make improvement for ssdk
Change mib polling task timer to 120 seconds
Use u64 data type for mib counter calculation
Setup correct HOL value for S17 chip
Remove vlan entry  with no member

+++ src/ref/ref_mib.c
@@ -72,9 +72,9 @@ _qca_ar8327_sw_capture_port_counter(struct switch_dev *dev, int port)
     priv->mib_counters[pos++] += mib_Info.RxMaxByte;
     priv->mib_counters[pos++] += mib_Info.RxTooLong;
     priv->mib_counters[pos] += mib_Info.RxGoodByte_lo;
-    priv->mib_counters[pos++] |= (mib_Info.RxGoodByte_hi << 32);
+    priv->mib_counters[pos++] |= (((u64)mib_Info.RxGoodByte_hi) << 32);
     priv->mib_counters[pos] += mib_Info.RxBadByte_lo;
-    priv->mib_counters[pos++] |= (mib_Info.RxBadByte_hi << 32);
+    priv->mib_counters[pos++] |= (((u64)mib_Info.RxBadByte_hi) << 32);
     priv->mib_counters[pos++] += mib_Info.RxOverFlow;
     priv->mib_counters[pos++] += mib_Info.Filtered;
     priv->mib_counters[pos++] += mib_Info.TxBroad;
@@ -90,7 +90,7 @@ _qca_ar8327_sw_capture_port_counter(struct switch_dev *dev, int port)
     priv->mib_counters[pos++] += mib_Info.TxMaxByte;
     priv->mib_counters[pos++] += mib_Info.TxOverSize;
     priv->mib_counters[pos] += mib_Info.TxByte_lo;
-    priv->mib_counters[pos++] |= (mib_Info.TxByte_hi << 32);
+    priv->mib_counters[pos++] |= (((u64)mib_Info.TxByte_hi) << 32);
     priv->mib_counters[pos++] += mib_Info.TxCollision;
     priv->mib_counters[pos++] += mib_Info.TxAbortCol;
     priv->mib_counters[pos++] += mib_Info.TxMultiCol;
@@ -130,9 +130,9 @@ _qca_ar8327_sw_capture_port_rx_counter(struct switch_dev *dev, int port)
     priv->mib_counters[pos++] += mib_Info.RxMaxByte;
     priv->mib_counters[pos++] += mib_Info.RxTooLong;
     priv->mib_counters[pos] += mib_Info.RxGoodByte_lo;
-    priv->mib_counters[pos++] |= (mib_Info.RxGoodByte_hi << 32);
+    priv->mib_counters[pos++] |= (((u64)mib_Info.RxGoodByte_hi) << 32);
     priv->mib_counters[pos] += mib_Info.RxBadByte_lo;
-    priv->mib_counters[pos++] |= (mib_Info.RxBadByte_hi << 32);
+    priv->mib_counters[pos++] |= (((u64)mib_Info.RxBadByte_hi) << 32);
     priv->mib_counters[pos++] += mib_Info.RxOverFlow;
     priv->mib_counters[pos++] += mib_Info.Filtered;
 
@@ -166,7 +166,7 @@ _qca_ar8327_sw_capture_port_tx_counter(struct switch_dev *dev, int port)
     priv->mib_counters[pos++] += mib_Info.TxMaxByte;
     priv->mib_counters[pos++] += mib_Info.TxOverSize;
     priv->mib_counters[pos] += mib_Info.TxByte_lo;
-    priv->mib_counters[pos++] |= (mib_Info.TxByte_hi << 32);
+    priv->mib_counters[pos++] |= (((u64)mib_Info.TxByte_hi) << 32);
     priv->mib_counters[pos++] += mib_Info.TxCollision;
     priv->mib_counters[pos++] += mib_Info.TxAbortCol;
     priv->mib_counters[pos++] += mib_Info.TxMultiCol;
